fix Dirichlet condition for AdDiffSEM example
the Dirichlet entry of the forcing term is the time derivative of the
Dirichlet data. In our example it had to be zero'ed out, since the
data was zero. Before, it was a very small nonzero number
corresponding to the current solution, which happend to be small due
to the initial condition and rightward advection.